backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
celery==5.3.6
redis==5.0.1
supabase==2.0.0
anthropic==0.21.3
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.2.0
numpy==1.26.3
httpx==0.24.1
